Break localization of map view into separate class.
