+++ app/src/main/java/org/opus/beacon/MapActivity.java
@@ -1,15 +1,10 @@
 package org.opus.beacon;
 
-import android.Manifest;
 import android.content.Context;
 import android.content.Intent;
-import android.content.pm.PackageManager;
 import android.graphics.Color;
 import android.location.Location;
-import android.location.LocationListener;
 import android.os.AsyncTask;
-import android.os.Handler;
-import android.support.v4.app.ActivityCompat;
 import android.os.Bundle;
 
 import com.google.android.gms.maps.GoogleMap;
@@ -18,8 +13,6 @@ import com.google.android.gms.maps.model.CameraPosition;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.Marker;
 import com.google.android.gms.maps.model.MarkerOptions;
-import android.location.LocationManager;
-import android.support.v4.content.ContextCompat;
 import android.util.Log;
 import android.view.View;
 import android.widget.TextView;
@@ -28,23 +21,15 @@ import android.widget.Toast;
 import java.util.ArrayList;
 import java.util.HashMap;
 
-public class MapActivity extends ConstrainedCameraMap
-    implements LocationListener,
-    ActivityCompat.OnRequestPermissionsResultCallback,
-    GoogleMap.OnMarkerClickListener {
 
-    private LocationManager mLocationManager;
+public class MapActivity extends LocalizingMap
+    implements GoogleMap.OnMarkerClickListener {
+
     private Context context;
-
-    private float GPS_ACCURACY = 50.0f; //meters
-    private long GPS_WAIT = 4000; //milliseconds
-
-    private int ACCESS_FINE_LOCATION_TAG = 125;
     private int BEACON_SUBMISSION = 37;
 
     private BeaconRestClient mClient;
     private Auth mAuth;
-
     private HashMap<Marker, BeaconThumb> mMarkerHash;
 
     @Override
@@ -62,30 +47,15 @@ public class MapActivity extends ConstrainedCameraMap
     }
 
     @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        if (requestCode == BEACON_SUBMISSION && resultCode == RESULT_OK) {
-           localize();
-        }
+    protected void setUpMap() {
+        super.setUpMap();
+        mMap.setOnMarkerClickListener(this);
     }
 
     @Override
-    protected void setUpMap() {
-        super.setUpMap();
-        mLocationManager = (LocationManager) getSystemService(Context.LOCATION_SERVICE);
-        mMap.setOnMarkerClickListener(this);
-        localize();
-    }
-
-    private void localize() {
-        if (ContextCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION) !=
-                PackageManager.PERMISSION_GRANTED) {
-
-            ActivityCompat.requestPermissions(this,
-                    new String[]{Manifest.permission.ACCESS_FINE_LOCATION},
-                    ACCESS_FINE_LOCATION_TAG);
-        } else {
-            mLocationManager.requestSingleUpdate(LocationManager.NETWORK_PROVIDER, this, null);
-            mLocationManager.requestLocationUpdates(mLocationManager.GPS_PROVIDER, 1000, 1, this);
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        if (requestCode == BEACON_SUBMISSION && resultCode == RESULT_OK) {
+           localize();
         }
     }
 
@@ -97,55 +67,8 @@ public class MapActivity extends ConstrainedCameraMap
         toast.show();
     }
 
-    private boolean shouldUseLocation(Location location) {
-        return location.getAccuracy() < GPS_ACCURACY;
-    }
-
-    private void performInitialZoom(Location location) {
-        zoomToLocation(location);
-        mLocationManager.removeUpdates(this);
+    public void execute(Location location) {
         new GetLocalBeacons().execute(location);
-    }
-
-    private Handler mLocalizationTimeoutHandler = null;
-
-    @Override
-    public void onLocationChanged(Location location) {
-        final Location loc = location;
-        if (mLocalizationTimeoutHandler == null) {
-            mLocalizationTimeoutHandler = new Handler();
-            mLocalizationTimeoutHandler.postDelayed( new Runnable() {
-                @Override
-                public void run() {
-                    performInitialZoom(loc);
-                }
-            }, GPS_WAIT);
-        }
-
-        if (!shouldUseLocation(location))
-            return;
-
-        performInitialZoom(location);
-        mLocalizationTimeoutHandler.removeCallbacksAndMessages(null);
-    }
-
-    @Override
-    public void onStatusChanged(String provider, int status, Bundle extras) {}
-
-    @Override
-    public void onProviderEnabled(String provider) {}
-
-    @Override
-    public void onProviderDisabled(String provider) {}
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, String permissions[], int[] grantResults) {
-        if (requestCode != ACCESS_FINE_LOCATION_TAG)
-            return;
-
-        if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-            localize();
-        }
     }
 
     @Override
